Extract classes from interface
