+++ fixture-monkey-junit5/src/main/java/com/navercorp/fixturemonkey/junit/jupiter/extension/support/ParameterContextAwareFixtureMonkey.java
@@ -18,8 +18,6 @@
 
 package com.navercorp.fixturemonkey.junit.jupiter.extension.support;
 
-import java.lang.reflect.Parameter;
-import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.List;
 import java.util.stream.Stream;
@@ -27,7 +25,6 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.extension.ParameterContext;
 
 import com.navercorp.fixturemonkey.FixtureMonkey;
-import com.navercorp.fixturemonkey.junit.jupiter.annotation.GiveMe;
 
 public interface ParameterContextAwareFixtureMonkey {
 
@@ -41,70 +38,6 @@ public interface ParameterContextAwareFixtureMonkey {
 			return new StreamParameterContextAwareFixtureMonkey(parameterContext, fixtureMonkey);
 		} else {
 			return new DefaultParameterContextAwareFixtureMonkey(parameterContext, fixtureMonkey);
-		}
-	}
-
-	class DefaultParameterContextAwareFixtureMonkey implements ParameterContextAwareFixtureMonkey {
-		private final ParameterContext parameterContext;
-		private final FixtureMonkey fixtureMonkey;
-
-		public DefaultParameterContextAwareFixtureMonkey(ParameterContext parameterContext,
-			FixtureMonkey fixtureMonkey) {
-			this.parameterContext = parameterContext;
-			this.fixtureMonkey = fixtureMonkey;
-		}
-
-		@Override
-		public Object giveMe() {
-			Class<?> type = parameterContext.getParameter().getType();
-
-			return fixtureMonkey.giveMeOne(type);
-		}
-	}
-
-	class ListParameterContextAwareFixtureMonkey implements ParameterContextAwareFixtureMonkey {
-		private final ParameterContext parameterContext;
-		private final FixtureMonkey fixtureMonkey;
-
-		public ListParameterContextAwareFixtureMonkey(ParameterContext parameterContext, FixtureMonkey fixtureMonkey) {
-			if (parameterContext.getParameter().getType() != List.class) {
-				throw new IllegalArgumentException("Type of parameter must be List.");
-			}
-			this.parameterContext = parameterContext;
-			this.fixtureMonkey = fixtureMonkey;
-		}
-
-		@Override
-		public List<?> giveMe() {
-			Parameter parameter = parameterContext.getParameter();
-			ParameterizedType parameterizedType = (ParameterizedType)parameter.getParameterizedType();
-			Class<?> genericType = (Class<?>)parameterizedType.getActualTypeArguments()[0];
-
-			int size = parameter.getAnnotation(GiveMe.class).size();
-			return fixtureMonkey.giveMe(genericType, size);
-		}
-	}
-
-	class StreamParameterContextAwareFixtureMonkey implements ParameterContextAwareFixtureMonkey {
-		private final ParameterContext parameterContext;
-		private final FixtureMonkey fixtureMonkey;
-
-		public StreamParameterContextAwareFixtureMonkey(ParameterContext parameterContext,
-			FixtureMonkey fixtureMonkey) {
-			if (parameterContext.getParameter().getType() != Stream.class) {
-				throw new IllegalArgumentException("Type of parameter must be Stream.");
-			}
-			this.parameterContext = parameterContext;
-			this.fixtureMonkey = fixtureMonkey;
-		}
-
-		@Override
-		public Stream<?> giveMe() {
-			Parameter parameter = parameterContext.getParameter();
-			ParameterizedType parameterizedType = (ParameterizedType)parameter.getParameterizedType();
-			Class<?> genericType = (Class<?>)parameterizedType.getActualTypeArguments()[0];
-
-			return fixtureMonkey.giveMe(genericType);
 		}
 	}
 }
